인증 로직 분리를 위한 AuthService 생성
